Nail-related preparations before introducing new format:  Utilize BRDPartMountingSide for the nails side marking  Move inverting Y axis for BRD2 format nails nearer to the reading point  Separate specific BRDFile implementation from common base class BRDFileBase  Factor out Nails-as-pins conversion to a method of BRDFileBase from specific BRD2 implementation.
