+++ src/openboardview/BoardView.h
@@ -123,7 +123,7 @@ enum DrawChannel {
 enum FlipModes { flipModeVP = 0, flipModeMP = 1, NUM_FLIP_MODES };
 
 struct BoardView {
-	BRDFile *m_file;
+	BRDFileBase *m_file;
 	Board *m_board;
 	BackgroundImage backgroundImage{m_current_side};
 
@@ -326,7 +326,7 @@ struct BoardView {
 	void DrawParts(ImDrawList *draw);
 	void DrawBoard();
 	void DrawNetWeb(ImDrawList *draw);
-	void LoadBoard(BRDFile *file);
+	void LoadBoard(BRDFileBase *file);
 	int LoadFile(const filesystem::path &filepath);
 	ImVec2 CoordToScreen(float x, float y, float w = 1.0f);
 	ImVec2 ScreenToCoord(float x, float y, float w = 1.0f);
