+++ src/openboardview/FileFormats/BRDFile.cpp
@@ -1,6 +1,5 @@
 #include "BRDFile.h"
 
-#include "utf8/utf8.h"
 #include "utils.h"
 #include <cctype>
 #include <stdexcept>
@@ -10,72 +9,6 @@
 
 // Header for recognizing a BRD file
 decltype(BRDFile::signature) constexpr BRDFile::signature;
-
-// from stb.h
-void stringfile(char *buffer, std::vector<char*> &lines) {
-	char *s;
-	size_t count, i;
-
-	// two passes through: first time count lines, second time set them
-	for (i = 0; i < 2; ++i) {
-		s                   = buffer;
-		if (i == 1) lines[0] = s;
-		count               = 1; // was '1',  but C arrays are 0-indexed
-		while (*s) {
-			if (*s == '\n' || *s == '\r') {
-
-				// If this is the 2nd pass, then terminate the line at the first line break char
-				if (i == 1) *s = 0;
-
-				s++; // next char
-
-				// if the termination is a CRLF combo, then jump to the next char
-				if ((*s == '\r') || (*s == '\n')) s++;
-
-				// if the char is valid (first after line break), set up the next item in the line array
-				if (*s) { // it's not over yet
-					if (i == 1) {
-						lines[count] = s;
-						//					  list[count+1] = NULL;
-						// fprintf(stdout,"%s\n",list[count]);
-					}
-					++count;
-				}
-			}
-			s++;
-		} // while s
-
-		// Generate the required array to hold all the line starting points
-		if (i == 0) {
-			lines.resize(count);
-		}
-	}
-}
-
-char *fix_to_utf8(char *s, char **arena, char *arena_end) {
-	if (!utf8valid(s)) {
-		return s;
-	}
-	char *p     = *arena;
-	char *begin = p;
-	while (*s) {
-		uint32_t c = (uint8_t)*s;
-		if (c < 0x80) {
-			if (p + 1 >= arena_end) goto done;
-			*p++ = c;
-		} else {
-			if (p + 2 >= arena_end) goto done;
-			*p++ = 0xc0 | (c >> 6);
-			*p++ = 0x80 | (c & 0x3f);
-		}
-		++s;
-	}
-	if (p + 1 >= arena_end) goto done;
-	*p++ = 0;
-done:
-	*arena = p;
-	return begin;
-}
 
 /*
  * Returns true if the file format seems to be BRD.
@@ -196,7 +129,7 @@ BRDFile::BRDFile(std::vector<char> &buf) {
 				nail.probe = READ_UINT();
 				nail.pos.x = READ_INT();
 				nail.pos.y = READ_INT();
-				nail.side  = READ_UINT();
+				nail.side  = READ_UINT() == 1 ? BRDPartMountingSide::Top : BRDPartMountingSide::Bottom;
 				nail.net   = READ_STR();
 				nails.push_back(nail);
 			} break;
