+++ src/openboardview/FileFormats/BVRFile.cpp
@@ -138,9 +138,9 @@ BVRFile::BVRFile(std::vector<char> &buf) {
 				/*char *grid =*/READ_STR();
 				char *loc = READ_STR();
 				if (!strcmp(loc, "(T)"))
-					nail.side = 1;
+					nail.side = BRDPartMountingSide::Top;
 				else
-					nail.side = 2;
+					nail.side = BRDPartMountingSide::Bottom;
 
 				/*char *netid =*/READ_STR();
 				nail.net   = READ_STR();
